Create day15.py
DP may be easier and faster. However, I wrote a maze generator yesterday and discovered that the problem of day 15 has some similarities with the maze solver. As a result, I am learning A* and trying to use it to solve the problem too.
fk covid19. I spent 4 hours doing a nucleic acid test, so I lost a green mark on my contribution board. TT